work item: 29772
git-svn-id: https://dbentry.svn.codeplex.com/svn@82005 81e6e737-5a20-4d0f-a97c-cb01297b46e0

+++ src/Lephone.UnitTest/Data/TableSchemaTest.cs
@@ -24,7 +24,7 @@ namespace Lephone.UnitTest.Data
 
             Assert.AreEqual("BelongsTo_Id", ls[2].ColumnName);
             Assert.AreEqual(false, ls[1].IsKey);
-            Assert.AreEqual(typeof(int), ls[2].DataType);
+            Assert.AreEqual(typeof(long), ls[2].DataType);
         }
     }
 }
